Created abstract view controller for passphrase creation; Fixed PR comments;
