+++ FlowCrypt/Controllers/Setup/SetupCreatePassphraseAbstractViewController.swift
@@ -1,8 +1,8 @@
 //
-//  CreatePrivateKeyViewController.swift
+//  SetupCreatePassphraseAbstractViewController.swift
 //  FlowCrypt
 //
-//  Created by Anton Kharchevskyi on 23.05.2021.
+//  Created by Yevhen Kyivskyi on 13.08.2021.
 //  Copyright © 2021 FlowCrypt Limited. All rights reserved.
 //
 
@@ -11,45 +11,26 @@ import FlowCryptCommon
 import FlowCryptUI
 import Promises
 
-enum CreateKeyError: Error {
-    case weakPassPhrase(_ strength: CoreRes.ZxcvbnStrengthBar)
-    // Missing user email
-    case missedUserEmail
-    // Missing user name
-    case missedUserName
-    // Pass phrases don't match
-    case doesntMatch
-    // silent abort
-    case conformingPassPhraseError
-}
-
-enum CreatePassphraseWithExistingKeyError: Error {
-    // No private key was found
-    case noPrivateKey
-}
-
 /**
- * Controller which is responsible for generating a new key during setup
- * - User is sent here from **SetupInitialViewController** in case there are no backups found
- * - Here user can enter a pass phrase (can be saved in memory or in encrypted storage) and generate a key
- * - After key is generated, user will be redirected to **main flow** (inbox view)
+ * Controller which decalres a base logic for passphrase setup
+ * - Has not to have an instance!
  */
 
-final class SetupCreatePassphraseViewController: TableNodeViewController, PassPhraseSaveable {
+class SetupCreatePassphraseAbstractViewController: TableNodeViewController, PassPhraseSaveable {
     enum Parts: Int, CaseIterable {
         case title, description, passPhrase, divider, saveLocally, saveInMemory, action, subtitle
     }
 
-    private let parts = Parts.allCases
-    private let decorator: SetupViewDecorator
-    private let core: Core
-    private let router: GlobalRouterType
-    private let user: UserId
-    private let backupService: BackupServiceType
-    private let storage: DataServiceType
-    private let keyStorage: KeyStorageType
-    private let attester: AttesterApiType
-    private let keys: [CoreRes.ParseKeys]
+    var parts: [Parts] {
+        Parts.allCases
+    }
+
+    let decorator: SetupViewDecorator
+    let core: Core
+    let router: GlobalRouterType
+    let user: UserId
+    let storage: DataServiceType
+    let keyStorage: KeyStorageType
     let passPhraseService: PassPhraseServiceType
 
     var shouldSaveLocally = true {
@@ -69,26 +50,20 @@ final class SetupCreatePassphraseViewController: TableNodeViewController, PassPh
 
     init(
         user: UserId,
-        keys: [CoreRes.ParseKeys] = [],
-        backupService: BackupServiceType = BackupService(),
         core: Core = .shared,
         router: GlobalRouterType = GlobalRouter(),
         decorator: SetupViewDecorator = SetupViewDecorator(),
         storage: DataServiceType = DataService.shared,
         keyStorage: KeyStorageType = KeyDataStorage(),
-        attester: AttesterApiType = AttesterApi(),
         passPhraseService: PassPhraseServiceType = PassPhraseService()
     ) {
         self.user = user
         self.core = core
         self.router = router
         self.decorator = decorator
-        self.backupService = backupService
         self.storage = storage
-        self.attester = attester
         self.keyStorage = keyStorage
         self.passPhraseService = passPhraseService
-        self.keys = keys
         super.init(node: TableNode())
     }
 
@@ -101,12 +76,16 @@ final class SetupCreatePassphraseViewController: TableNodeViewController, PassPh
         super.viewDidLoad()
         setupUI()
     }
+
+    func setupAccount(with passphrase: String) {
+        fatalError("This method has to be overriden")
+    }
 }
 
 // MARK: - UI
 
-extension SetupCreatePassphraseViewController {
-    private func setupUI() {
+extension SetupCreatePassphraseAbstractViewController {
+    func setupUI() {
         node.delegate = self
         node.dataSource = self
 
@@ -144,110 +123,9 @@ extension SetupCreatePassphraseViewController {
 
 // MARK: - Setup
 
-extension SetupCreatePassphraseViewController {
-    private func setupAccountWithGeneratedKey(with passPhrase: String) {
-        Promise { [weak self] in
-            guard let self = self else { return }
-            self.showSpinner()
+extension SetupCreatePassphraseAbstractViewController {
 
-            let userId = try self.getUserId()
-
-            try awaitPromise(self.validateAndConfirmNewPassPhraseOrReject(passPhrase: passPhrase))
-
-            let encryptedPrv = try self.core.generateKey(passphrase: passPhrase, variant: .curve25519, userIds: [userId])
-
-            try awaitPromise(self.backupService.backupToInbox(keys: [encryptedPrv.key], for: self.user))
-
-            let passPhrase = PassPhrase(value: passPhrase, fingerprints: encryptedPrv.key.fingerprints)
-
-            self.keyStorage.addKeys(keyDetails: [encryptedPrv.key], source: .generated, for: self.user.email)
-            self.passPhraseService.savePassPhrase(with: passPhrase, inStorage: self.shouldSaveLocally)
-
-            let updateKey = self.attester.updateKey(
-                email: userId.email,
-                pubkey: encryptedPrv.key.public,
-                token: self.storage.token
-            )
-
-            try awaitPromise(self.alertAndSkipOnRejection(
-                updateKey,
-                fail: "Failed to submit Public Key")
-            )
-            let testWelcome = self.attester.testWelcome(email: userId.email, pubkey: encryptedPrv.key.public)
-            try awaitPromise(self.alertAndSkipOnRejection(
-                testWelcome,
-                fail: "Failed to send you welcome email")
-            )
-        }
-        .then(on: .main) { [weak self] in
-            self?.hideSpinner()
-            self?.moveToMainFlow()
-        }
-        .catch(on: .main) { [weak self] error in
-            guard let self = self else { return }
-            self.hideSpinner()
-
-            let isErrorHandled = self.handleCommon(error: error)
-
-            if !isErrorHandled {
-                self.showAlert(error: error, message: "Could not finish setup, please try again")
-            }
-        }
-    }
-
-    private func setupAccountWithExistingKeys(with passPhrase: String) {
-        Promise { [weak self] in
-            guard let self = self else { return }
-            self.showSpinner()
-
-            try awaitPromise(self.validateAndConfirmNewPassPhraseOrReject(passPhrase: passPhrase))
-
-            try self.keys.forEach { key in
-                try key.keyDetails.forEach { keyDetail in
-                    guard let privateKey = keyDetail.private else {
-                        throw CreatePassphraseWithExistingKeyError.noPrivateKey
-                    }
-                    let encryptedPrv = try self.core.encryptKey(
-                        armoredPrv: privateKey,
-                        passphrase: passPhrase
-                    )
-                    let parsedKey = try self.core.parseKeys(armoredOrBinary: encryptedPrv.encryptedKey.data())
-                    parsedKey.keyDetails.forEach { keyDetails in
-                        let passPhrase = PassPhrase(value: passPhrase, fingerprints: keyDetails.fingerprints)
-                        self.passPhraseService.savePassPhrase(with: passPhrase, inStorage: self.shouldSaveLocally)
-                    }
-                    try awaitPromise(self.backupService.backupToInbox(keys: parsedKey.keyDetails, for: self.user))
-                    self.keyStorage.addKeys(keyDetails: parsedKey.keyDetails, source: .generated, for: self.user.email)
-                }
-            }
-        }
-        .then(on: .main) { [weak self] in
-            self?.hideSpinner()
-            self?.moveToMainFlow()
-        }
-        .catch(on: .main) { [weak self] error in
-            guard let self = self else { return }
-            self.hideSpinner()
-
-            let isErrorHandled = self.handleCommon(error: error)
-
-            if !isErrorHandled {
-                self.showAlert(error: error, message: "Could not finish setup, please try again")
-            }
-        }
-    }
-
-    private func getUserId() throws -> UserId {
-        guard let email = DataService.shared.email, !email.isEmpty else {
-            throw CreateKeyError.missedUserEmail
-        }
-        guard let name = DataService.shared.email, !name.isEmpty else {
-            throw CreateKeyError.missedUserName
-        }
-        return UserId(email: email, name: name)
-    }
-
-    private func validateAndConfirmNewPassPhraseOrReject(passPhrase: String) -> Promise<Void> {
+    func validateAndConfirmNewPassPhraseOrReject(passPhrase: String) -> Promise<Void> {
         Promise { [weak self] in
             guard let self = self else { throw AppErr.nilSelf }
 
@@ -296,8 +174,8 @@ extension SetupCreatePassphraseViewController {
     }
 }
 
-extension SetupCreatePassphraseViewController {
-    private func moveToMainFlow() {
+extension SetupCreatePassphraseAbstractViewController {
+    func moveToMainFlow() {
         router.proceed()
     }
 
@@ -312,24 +190,21 @@ extension SetupCreatePassphraseViewController {
             return
         }
         logger.logInfo("Setup account with \(passPhrase)")
-        if keys.isNotEmpty {
-            self.setupAccountWithExistingKeys(with: passPhrase)
-        } else {
-            setupAccountWithGeneratedKey(with: passPhrase)
-        }
+        setupAccount(with: passPhrase)
     }
 }
 
 // MARK: - ASTableDelegate, ASTableDataSource
 
-extension SetupCreatePassphraseViewController: ASTableDelegate, ASTableDataSource {
+extension SetupCreatePassphraseAbstractViewController: ASTableDelegate, ASTableDataSource {
     func tableNode(_: ASTableNode, numberOfRowsInSection _: Int) -> Int {
         parts.count
     }
 
     func tableNode(_: ASTableNode, nodeBlockForRowAt indexPath: IndexPath) -> ASCellNodeBlock {
         return { [weak self] in
-            guard let self = self, let part = Parts(rawValue: indexPath.row) else { return ASCellNode() }
+            guard let self = self else { return ASCellNode() }
+            let part = self.parts[indexPath.row]
             switch part {
             case .title:
                 return SetupTitleNode(
@@ -387,7 +262,7 @@ extension SetupCreatePassphraseViewController: ASTableDelegate, ASTableDataSourc
     }
 
     func tableNode(_ tableNode: ASTableNode, didSelectRowAt indexPath: IndexPath) {
-        guard let part = Parts(rawValue: indexPath.row) else { return }
+        let part = parts[indexPath.row]
 
         switch part {
         case .description:
